Refactor workers/ models to new bokeh interface
Also refactor tests a bit
