+++ distributed/bokeh/components.py
@@ -10,6 +10,9 @@ from bokeh.models import (
     PanTool, WheelZoomTool, Title, Range1d, Quad, Text, value, Line,
     NumeralTickFormatter, ToolbarBox, Legend, LegendItem
 )
+from bokeh.models.widgets import DataTable, TableColumn, NumberFormatter
+from bokeh.palettes import Spectral9
+from bokeh.plotting import figure
 
 from distributed.diagnostics.progress_stream import progress_quads, nbytes_bar
 from distributed.utils import log_errors
@@ -354,3 +357,189 @@ class ResourceProfiles(DashboardComponent):
             data = {k: [v[i] for i in indexes] for k, v in data.items()}
             self.resource_index[0] = index[-1]
             self.source.stream(data, 1000)
+
+
+class WorkerTable(DashboardComponent):
+    """ Column data source and plot for host table """
+    def __init__(self, **kwargs):
+        # names = ['host', 'cpu', 'memory_percent', 'memory', 'cores', 'processes',
+        #          'processing', 'latency', 'last-seen', 'disk-read', 'disk-write',
+        #          'network-send', 'network-recv']
+        with log_errors():
+            names = ['processes', 'disk-read', 'cores', 'cpu', 'disk-write',
+                     'memory', 'last-seen', 'memory_percent', 'host']
+            self.source = ColumnDataSource({k: [] for k in names})
+
+            columns = {name: TableColumn(field=name,
+                                         title=name.replace('_percent', ' %'))
+                       for name in names}
+
+            cnames = ['host', 'cores', 'processes', 'memory', 'cpu', 'memory_percent']
+
+            formatters = {'cpu': NumberFormatter(format='0.0 %'),
+                          'memory_percent': NumberFormatter(format='0.0 %'),
+                          'memory': NumberFormatter(format='0 b'),
+                          'latency': NumberFormatter(format='0.00000'),
+                          'last-seen': NumberFormatter(format='0.000'),
+                          'disk-read': NumberFormatter(format='0 b'),
+                          'disk-write': NumberFormatter(format='0 b'),
+                          'net-send': NumberFormatter(format='0 b'),
+                          'net-recv': NumberFormatter(format='0 b')}
+
+            table = DataTable(source=self.source, columns=[columns[n] for n in cnames],
+                              id='bk-worker-table', **kwargs)
+            for name in cnames:
+                if name in formatters:
+                    table.columns[cnames.index(name)].formatter = formatters[name]
+
+            x_range = Range1d(0, 1)
+            y_range = Range1d(-.1, .1)
+            mem_plot = figure(title="Memory Usage (%)", tools='box_select',
+                              height=90, width=600, x_range=x_range, y_range=y_range,
+                              toolbar_location=None, id='bk-memory-usage-plot')
+            mem_plot.circle(source=self.source, x='memory_percent', y=0, size=10,
+                            alpha=0.5)
+            mem_plot.yaxis.visible = False
+            mem_plot.xaxis.minor_tick_line_width = 0
+            mem_plot.ygrid.visible = False
+            mem_plot.xaxis.minor_tick_line_alpha = 0
+
+            hover = HoverTool()
+            mem_plot.add_tools(hover)
+            hover = mem_plot.select(HoverTool)
+            hover.tooltips = """
+            <div>
+              <span style="font-size: 10px; font-family: Monaco, monospace;">@host: </span>
+              <span style="font-size: 10px; font-family: Monaco, monospace;">@memory_percent</span>
+            </div>
+            """
+            hover.point_policy = 'follow_mouse'
+
+            if 'sizing_mode' in kwargs:
+                sizing_mode = {'sizing_mode': kwargs['sizing_mode']}
+            else:
+                sizing_mode = {}
+
+            self.root = column(mem_plot, table, **sizing_mode)
+
+    def update(self, messages):
+        with log_errors():
+            try:
+                d = messages['workers']['deque'][-1]
+            except IndexError:
+                return
+
+            workers = sorted(d)
+
+            data = {}
+            data['host'] = workers
+            for name in ['cores', 'cpu', 'memory_percent', 'latency', 'last-seen',
+                         'memory', 'disk-read', 'disk-write', 'net-send',
+                         'net-recv']:
+                try:
+                    if name in ('cpu', 'memory_percent'):
+                        data[name] = [d[w][name] / 100 for w in workers]
+                    else:
+                        data[name] = [d[w][name] for w in workers]
+                except KeyError:
+                    pass
+
+            data['processing'] = [sorted(d[w]['processing']) for w in workers]
+            data['processes'] = [len(d[w]['ports']) for w in workers]
+            self.source.data.update(data)
+
+
+class ProcessingStacks(DashboardComponent):
+    def __init__(self, **kwargs):
+        with log_errors():
+            data = self.processing_update({'processing': {}, 'stacks': {},
+                                      'ncores': {}, 'ready': 0})
+            self.source = ColumnDataSource(data)
+
+            x_range = Range1d(-1, 1)
+            fig = figure(title='Processing and Pending', tools='resize',
+                         x_range=x_range, id='bk-processing-plot', **kwargs)
+            fig.quad(source=self.source, left=0, right='right', color=Spectral9[0],
+                     top='top', bottom='bottom')
+            fig.quad(source=self.source, left='left', right=0, color=Spectral9[1],
+                     top='top', bottom='bottom', alpha='alpha')
+
+            fig.xaxis.minor_tick_line_alpha = 0
+            fig.yaxis.visible = False
+            fig.ygrid.visible = False
+
+            hover = HoverTool()
+            fig.add_tools(hover)
+            hover = fig.select(HoverTool)
+            hover.tooltips = """
+            <div>
+                <span style="font-size: 14px; font-weight: bold;">Host:</span>&nbsp;
+                <span style="font-size: 10px; font-family: Monaco, monospace;">@name</span>
+            </div>
+            <div>
+                <span style="font-size: 14px; font-weight: bold;">Stacks:</span>&nbsp;
+                <span style="font-size: 10px; font-family: Monaco, monospace;">@stacks</span>
+            </div>
+            <div>
+                <span style="font-size: 14px; font-weight: bold;">Processing:</span>&nbsp;
+                <span style="font-size: 10px; font-family: Monaco, monospace;">@processing</span>
+            </div>
+            """
+            hover.point_policy = 'follow_mouse'
+
+            self.root = fig
+
+    def update(self, messages):
+        with log_errors():
+            msg = messages['processing']
+            if not msg.get('ncores'):
+                return
+            data = self.processing_update(msg)
+            x_range = self.root.x_range
+            max_right = max(data['right'])
+            min_left = min(data['left'][:-1])
+            cores = max(data['ncores'])
+            if min_left < x_range.start:  # not out there enough, jump ahead
+                x_range.start = min_left - 2
+            elif x_range.start < 2 * min_left - cores:  # way out there, walk back
+                x_range.start = x_range.start * 0.95 + min_left * 0.05
+            if x_range.end < max_right:
+                x_range.end = max_right + 2
+            elif x_range.end > 2 * max_right + cores:  # way out there, walk back
+                x_range.end = x_range.end * 0.95 + max_right * 0.05
+
+            self.source.data.update(data)
+
+    @staticmethod
+    def processing_update(msg):
+        with log_errors():
+            names = sorted(msg['stacks'])
+            stacks = msg['stacks']
+            stacks = [stacks[name] for name in names]
+            names = sorted(names)
+            processing = msg['processing']
+            processing = [processing[name] for name in names]
+            ncores = msg['ncores']
+            ncores = [ncores[name] for name in names]
+            n = len(names)
+            d = {'name': list(names),
+                 'processing': processing,
+                 'stacks': list(stacks),
+                 'right': list(processing),
+                 'left': [-s for s in stacks],
+                 'top': list(range(n, 0, -1)),
+                 'bottom': list(range(n - 1, -1, -1)),
+                 'ncores': ncores}
+
+            d['name'].append('ready')
+            d['processing'].append(0)
+            d['stacks'].append(msg['ready'])
+            d['left'].append(-msg['ready'] / n if n else 0)
+            d['right'].append(0)
+            d['top'].append(n)
+            d['bottom'].append(0)
+            d['ncores'].append(sum(msg['ncores'].values()))
+
+            d['alpha'] = [0.7] * n + [0.2]
+
+            return d
