+++ distributed/bokeh/tasks/server_lifecycle.py
@@ -69,12 +69,4 @@ def task_events(interval, deque, times, index, rectangles, workers, last_seen):
 n = 100000
 
 def on_server_loaded(server_context):
-    messages['task-events'] = {'interval': 200,
-                               'deque': deque(maxlen=n),
-                               'times': deque(maxlen=n),
-                               'index': deque(maxlen=n),
-                               'rectangles':{name: deque(maxlen=n) for name in
-                                            'start duration key name color worker worker_thread y alpha'.split()},
-                               'workers': dict(),
-                               'last_seen': [time()]}
     IOLoop.current().add_callback(task_events, **messages['task-events'])
