+++ distributed/bokeh/tests/test_worker_monitor.py
@@ -3,50 +3,10 @@ from __future__ import print_function, division, absolute_import
 import pytest
 pytest.importorskip('bokeh')
 
-from collections import deque
-import datetime
-
 from tornado import gen
 
-from distributed.bokeh.worker_monitor import (resource_profile_plot,
-        resource_profile_update, resource_append, worker_table_plot,
-        worker_table_update)
-from distributed.diagnostics.scheduler import workers, tasks
+from distributed.bokeh.worker_monitor import resource_append
 from distributed.utils_test import gen_cluster
-
-
-@gen_cluster()
-def test_resource_monitor_plot(s, a, b):
-    while any('last-seen' not in v for v in s.host_info.values()):
-        yield gen.sleep(0.01)
-
-    times_buffer = [1000, 1001, 1003]
-    workers_buffer = [{},
-                      {'10.10.20.86': {'cpu': 15.9, 'memory_percent': 63.0}},
-                      {'10.10.20.86': {'cpu': 14.9, 'memory_percent': 64.0,
-                                       'network-send': 2**16,
-                                       'network-recv': 2**15},
-                       '10.10.20.87': {'cpu': 13.9, 'memory_percent': 64.0,
-                                       'network-send': 2**17,
-                                       'network-recv': 2**16}}]
-
-    source, _, _, _ = resource_profile_plot()
-    resource_profile_update(source, workers_buffer, times_buffer)
-
-    assert source.data['workers'] == ['10.10.20.86', '10.10.20.87']
-    assert len(source.data['cpu']) == 2
-    assert source.data['cpu'][0] == ['null', 15.9, 14.9]
-    assert source.data['cpu'][1] == ['null', 'null', 13.9]
-
-    assert source.data['times'][0] == ['null', 1001000, 1003000]
-    assert source.data['times'][1] == ['null','null', 1003000]
-    assert len(source.data['times']) == 2
-    assert len(source.data['memory_percent']) == 2
-
-    assert len(source.data['network-send']) == 2
-    assert source.data['network-send'][1] == ['null', 'null', 2**17]
-    assert len(source.data['network-recv']) == 2
-    assert source.data['network-recv'][1] == ['null', 'null', 2**16]
 
 
 def test_resource_append():
@@ -62,20 +22,9 @@ def test_resource_append():
                      'network-send': [0.1875], 'network-recv': [0.09375]}
 
 
-@gen_cluster()
-def test_worker_table(s, a, b):
-    while any('last-seen' not in v for v in s.host_info.values()):
-        yield gen.sleep(0.01)
-    data = workers(s)
-    source, plot = worker_table_plot()
-    worker_table_update(source, data)
-
-    assert source.data['host'] == ['127.0.0.1']
-
-
 def test_processing_update():
     from distributed.diagnostics.scheduler import processing
-    from distributed.bokeh.worker_monitor import processing_update
+    from distributed.bokeh.components import ProcessingStacks
 
     class C(object):
         pass
@@ -97,7 +46,7 @@ def test_processing_update():
                    'memory': 1,
                    'ncores': {'alice': 4, 'bob': 4}}
 
-    data = processing_update(msg)
+    data = ProcessingStacks.processing_update(msg)
     expected = {'name': ['alice', 'bob', 'ready'],
                 'processing': [2, 1, 0],
                 'stacks': [3, 2, 7],
@@ -109,8 +58,3 @@ def test_processing_update():
                 'alpha': [0.7, 0.7, 0.2]}
 
     assert data == expected
-
-
-def test_processing_plot():
-    from distributed.bokeh.worker_monitor import processing_plot
-    source, fig = processing_plot()
