+++ distributed/bokeh/worker_monitor.py
@@ -78,24 +78,6 @@ def resource_profile_plot(sizing_mode='fixed', **kwargs):
     return source, p1, p2, combo_toolbar
 
 
-def resource_profile_update(source, worker_buffer, times_buffer):
-    data = defaultdict(list)
-
-    workers = sorted(list(set(chain(*list(w.keys() for w in worker_buffer)))))
-
-    for name in ['cpu', 'memory_percent', 'network-send', 'network-recv']:
-        data[name] = [[msg[w][name] if w in msg and name in msg[w] else 'null'
-                       for msg in worker_buffer]
-                       for w in workers]
-
-    data['workers'] = workers
-    data['times'] = [[t * 1000 if w in worker_buffer[i] else 'null'
-                      for i, t in enumerate(times_buffer)]
-                      for w in workers]
-
-    source.data.update(data)
-
-
 def resource_append(lists, msg):
     L = list(msg.values())
     if not L:
@@ -122,154 +104,3 @@ def resource_append(lists, msg):
 def mean(seq):
     seq = list(seq)
     return sum(seq) / len(seq)
-
-
-def worker_table_plot(**kwargs):
-    """ Column data source and plot for host table """
-    with log_errors():
-        # names = ['host', 'cpu', 'memory_percent', 'memory', 'cores', 'processes',
-        #          'processing', 'latency', 'last-seen', 'disk-read', 'disk-write',
-        #          'network-send', 'network-recv']
-        names = ['processes', 'disk-read', 'cores', 'cpu', 'disk-write',
-                 'memory', 'last-seen', 'memory_percent', 'host']
-        source = ColumnDataSource({k: [] for k in names})
-
-        columns = {name: TableColumn(field=name,
-                                     title=name.replace('_percent', ' %'))
-                   for name in names}
-
-        cnames = ['host', 'cores', 'processes', 'memory', 'cpu', 'memory_percent']
-
-        formatters = {'cpu': NumberFormatter(format='0.0 %'),
-                      'memory_percent': NumberFormatter(format='0.0 %'),
-                      'memory': NumberFormatter(format='0 b'),
-                      'latency': NumberFormatter(format='0.00000'),
-                      'last-seen': NumberFormatter(format='0.000'),
-                      'disk-read': NumberFormatter(format='0 b'),
-                      'disk-write': NumberFormatter(format='0 b'),
-                      'net-send': NumberFormatter(format='0 b'),
-                      'net-recv': NumberFormatter(format='0 b')}
-
-        table = DataTable(source=source, columns=[columns[n] for n in cnames],
-                          id='bk-worker-table', **kwargs)
-        for name in cnames:
-            if name in formatters:
-                table.columns[cnames.index(name)].formatter = formatters[name]
-
-        x_range = Range1d(0, 1)
-        y_range = Range1d(-.1, .1)
-        mem_plot = figure(title="Memory Usage (%)", tools='box_select',
-                          height=90, width=600, x_range=x_range, y_range=y_range,
-                          toolbar_location=None, id='bk-memory-usage-plot')
-        mem_plot.circle(source=source, x='memory_percent', y=0, size=10,
-                        alpha=0.5)
-        mem_plot.yaxis.visible = False
-        mem_plot.xaxis.minor_tick_line_width = 0
-        mem_plot.ygrid.visible = False
-        mem_plot.xaxis.minor_tick_line_alpha = 0
-
-        hover = HoverTool()
-        mem_plot.add_tools(hover)
-        hover = mem_plot.select(HoverTool)
-        hover.tooltips = """
-        <div>
-          <span style="font-size: 10px; font-family: Monaco, monospace;">@host: </span>
-          <span style="font-size: 10px; font-family: Monaco, monospace;">@memory_percent</span>
-        </div>
-        """
-        hover.point_policy = 'follow_mouse'
-
-    return source, [mem_plot, table]
-
-
-def worker_table_update(source, d):
-    """ Update host table source """
-    workers = sorted(d)
-
-    data = {}
-    data['host'] = workers
-    for name in ['cores', 'cpu', 'memory_percent', 'latency', 'last-seen',
-                 'memory', 'disk-read', 'disk-write', 'net-send',
-                 'net-recv']:
-        try:
-            if name in ('cpu', 'memory_percent'):
-                data[name] = [d[w][name] / 100 for w in workers]
-            else:
-                data[name] = [d[w][name] for w in workers]
-        except KeyError:
-            pass
-
-    data['processing'] = [sorted(d[w]['processing']) for w in workers]
-    data['processes'] = [len(d[w]['ports']) for w in workers]
-    source.data.update(data)
-
-
-def processing_plot(**kwargs):
-    data = processing_update({'processing': {}, 'stacks': {}, 'ncores': {},
-        'ready': 0})
-    source = ColumnDataSource(data)
-
-    x_range = Range1d(-1, 1)
-    fig = figure(title='Processing and Pending', tools='resize',
-                 x_range=x_range, id='bk-processing-plot', **kwargs)
-    fig.quad(source=source, left=0, right='right', color=Spectral9[0],
-             top='top', bottom='bottom')
-    fig.quad(source=source, left='left', right=0, color=Spectral9[1],
-             top='top', bottom='bottom', alpha='alpha')
-
-    fig.xaxis.minor_tick_line_alpha = 0
-    fig.yaxis.visible = False
-    fig.ygrid.visible = False
-
-    hover = HoverTool()
-    fig.add_tools(hover)
-    hover = fig.select(HoverTool)
-    hover.tooltips = """
-    <div>
-        <span style="font-size: 14px; font-weight: bold;">Host:</span>&nbsp;
-        <span style="font-size: 10px; font-family: Monaco, monospace;">@name</span>
-    </div>
-    <div>
-        <span style="font-size: 14px; font-weight: bold;">Stacks:</span>&nbsp;
-        <span style="font-size: 10px; font-family: Monaco, monospace;">@stacks</span>
-    </div>
-    <div>
-        <span style="font-size: 14px; font-weight: bold;">Processing:</span>&nbsp;
-        <span style="font-size: 10px; font-family: Monaco, monospace;">@processing</span>
-    </div>
-    """
-    hover.point_policy = 'follow_mouse'
-    return source, fig
-
-
-def processing_update(msg):
-    names = sorted(msg['stacks'])
-    stacks = msg['stacks']
-    stacks = [stacks[name] for name in names]
-    names = sorted(names)
-    processing = msg['processing']
-    processing = [processing[name] for name in names]
-    ncores = msg['ncores']
-    ncores = [ncores[name] for name in names]
-    n = len(names)
-    d = {'name': list(names),
-         'processing': processing,
-         'stacks': list(stacks),
-         'right': list(processing),
-         'left': [-s for s in stacks],
-         'top': list(range(n, 0, -1)),
-         'bottom': list(range(n - 1, -1, -1)),
-         'ncores': ncores}
-
-    d['name'].append('ready')
-    d['processing'].append(0)
-    d['stacks'].append(msg['ready'])
-    d['left'].append(-msg['ready'] / n if n else 0)
-    d['right'].append(0)
-    d['top'].append(n)
-    d['bottom'].append(0)
-    d['ncores'].append(sum(msg['ncores'].values()))
-
-    d['alpha'] = [0.7] * n + [0.2]
-
-    return d
